File tree reshuffle
