+++ .flake8
@@ -1,0 +1,3 @@
+[flake8]
+extend-ignore = E203
+max-line-length = 88
