+++ examples/examples.py
@@ -1,9 +1,14 @@
-from populations import *
+import os
+import sys
+sys.path.append(os.path.join(os.path.dirname(__file__),'../'))
+
+from models.populations import IndexedPopulation, Population
+from models.variables import Parameter
 
 FlyEggs1 = Population("flyeggs1")
 FlyEggs2 = Population("flyeggs2")
 
-FlyEggs = Indexed_Population("flyeggs", (2,))
+FlyEggs = IndexedPopulation("flyeggs", (2,))
 
 FlyEggs.add_population(FlyEggs1, 0)
 FlyEggs.add_population(FlyEggs2, 1)
@@ -22,9 +27,11 @@ print(Flies.variables.get_symbols())
 
 Flies._add_parameter("basic", "growth_rate", "r", 45)
 
-Flies._add_update_rule(FlyEggs1.variable, 'r_growth_rate * x_flyeggs1')
+Flies._add_update_rule(FlyEggs1.variable, "r_growth_rate * x_flyeggs1")
 
-Flies._add_update_rule('x_flyeggs1', Flies.parameters.growth_rate.symbol * FlyEggs1.variable.symbol)
+Flies._add_update_rule(
+    "x_flyeggs1", Flies.parameters.growth_rate.symbol * FlyEggs1.variable.symbol
+)
 
 P = Parameter.basic("P", "param", "p")
 Q = Parameter.basic("Q", "param", "q")
@@ -34,6 +41,6 @@ X = Parameter.composite("X", "comp", P.symbol + Q.symbol**2)
 
 Y = Parameter.composite("Y", "comp", X.symbol * R.symbol)
 
-Z = Parameter.composite("Z", "comp", 'exp(param_P * comp_X + comp_Y)')
+Z = Parameter.composite("Z", "comp", "exp(param_P * comp_X + comp_Y)")
 
 print(str(Z))
